Use indicator icon to convey some information
The icon now changes to show whether a timer is in progress---a
Pomodoro or a break---, or whether the app is paused.  The Pomodorino
icon is displayed in other cases.

This patch also refactors how the AppIndicator label is set, but I
haven’t observed a change in how it behaves, namely, the label doesn’t
seem to show up.  Requires testing on multiple desktops, I suppose.

This is a partial implementation of #21, but the icons should be
replaced with Pomodorino-specific icons so that it’s easier to tell
it’s Pomodorino’s state in a potentially crowded systray.